TradingBot-main/requirements.txt
mistune
yfinance
pandas
mplfinance
scipy
google-generativeai
python-dotenv